py-electric-rhombus new feature
Cursor at pipe-symbol in column 4 after RET 

#! /usr/bin/env python
# -\*- coding: utf-8 -\*-

if __name__ == '__main__':
    from foo import \*
    foo([\"baz\"]) # list of foo's development directories

# limitations:
#
#   Some comments on limitations: <RET>
----|---------------------------

This was considered as a bug, as newlines in comments
should follow indentation. Which is okay, if commenting
continues, but not, if code follows. When newline is
inserted, Emacs can't know which kind of text will
follow: `py-electric-rhombus' makes
comment-honoring-indentation occur, if comment symbol
get's inserted.