Declarations - new edit functions
Deal with assigments resp. statements in current level
which don't open blocks. Provides common edit functions as
copy, mark, kill, goto beg/end.

Functions will be used typically to grasp
initialisations resp. assignements of variables between
the definition of a class or method and it's body.

Likewise to deal with global stuff at the head of a
file.